Remove test_setitem.py
