+++ python/cudf/cudf/tests/dataframe/indexing/test_setitem.py
@@ -1,12 +1,111 @@
 # Copyright (c) 2025, NVIDIA CORPORATION.
 
-
+import numpy as np
 import pandas as pd
+import pytest
 
-from cudf import DataFrame
+import cudf
+from cudf.testing import assert_eq
+
+
+@pytest.mark.parametrize("arg", [[True, False, True], [True, True, True]])
+@pytest.mark.parametrize("value", [0, -1])
+def test_dataframe_setitem_bool_mask_scaler(arg, value):
+    df = pd.DataFrame({"a": [1, 2, 3]})
+    gdf = cudf.from_pandas(df)
+
+    df[arg] = value
+    gdf[arg] = value
+    assert_eq(df, gdf)
+
+
+def test_dataframe_setitem_scaler_bool():
+    df = pd.DataFrame({"a": [1, 2, 3]})
+    df[[True, False, True]] = pd.DataFrame({"a": [-1, -2]})
+
+    gdf = cudf.DataFrame({"a": [1, 2, 3]})
+    gdf[[True, False, True]] = cudf.DataFrame({"a": [-1, -2]})
+    assert_eq(df, gdf)
+
+
+@pytest.mark.parametrize(
+    "df",
+    [pd.DataFrame({"a": [1, 2, 3]}), pd.DataFrame({"a": ["x", "y", "z"]})],
+)
+@pytest.mark.parametrize("arg", [["a"], "a", "b"])
+@pytest.mark.parametrize(
+    "value", [-10, pd.DataFrame({"a": [-1, -2, -3]}), "abc"]
+)
+def test_dataframe_setitem_columns(df, arg, value):
+    gdf = cudf.from_pandas(df)
+    cudf_replace_value = value
+
+    if isinstance(cudf_replace_value, pd.DataFrame):
+        cudf_replace_value = cudf.from_pandas(value)
+
+    df[arg] = value
+    gdf[arg] = cudf_replace_value
+    assert_eq(df, gdf, check_dtype=False)
+
+
+@pytest.mark.parametrize(
+    "value",
+    [
+        pd.DataFrame({"0": [-1, -2, -3], "1": [-0, -10, -1]}),
+        10,
+        "rapids",
+        0.32234,
+        np.datetime64(1324232423423342, "ns"),
+        np.timedelta64(34234324234324234, "ns"),
+    ],
+)
+def test_dataframe_setitem_new_columns(value):
+    df = pd.DataFrame({"a": [1, 2, 3]})
+    arg = ["b", "c"]
+    gdf = cudf.from_pandas(df)
+    cudf_replace_value = value
+
+    if isinstance(cudf_replace_value, pd.DataFrame):
+        cudf_replace_value = cudf.from_pandas(value)
+
+    df[arg] = value
+    gdf[arg] = cudf_replace_value
+    assert_eq(df, gdf, check_dtype=True)
+
+
+def test_series_setitem_index():
+    df = pd.DataFrame(
+        data={"b": [-1, -2, -3], "c": [1, 2, 3]}, index=[1, 2, 3]
+    )
+
+    df["b"] = pd.Series(data=[12, 11, 10], index=[3, 2, 1])
+    gdf = cudf.DataFrame(
+        data={"b": [-1, -2, -3], "c": [1, 2, 3]}, index=[1, 2, 3]
+    )
+    gdf["b"] = cudf.Series(data=[12, 11, 10], index=[3, 2, 1])
+    assert_eq(df, gdf, check_dtype=False)
+
+
+@pytest.mark.xfail(reason="Copy-on-Write should make a copy")
+@pytest.mark.parametrize(
+    "index",
+    [
+        pd.MultiIndex.from_frame(
+            pd.DataFrame({"b": [3, 2, 1], "c": ["a", "b", "c"]})
+        ),
+        ["a", "b", "c"],
+    ],
+)
+def test_setitem_dataframe_series_inplace(index):
+    gdf = cudf.DataFrame({"a": [1, 2, 3]}, index=index)
+    expected = gdf.copy()
+    with cudf.option_context("copy_on_write", True):
+        gdf["a"].replace(1, 500, inplace=True)
+
+    assert_eq(expected, gdf)
 
 
 def test_setitem_datetime():
-    df = DataFrame()
+    df = cudf.DataFrame()
     df["date"] = pd.date_range("20010101", "20010105").values
     assert df.date.dtype.kind == "M"
