+++ python/cudf/cudf/tests/series/indexing/test_iloc.py
@@ -1,9 +1,20 @@
 # Copyright (c) 2025, NVIDIA CORPORATION.
 
+import numpy as np
 import pytest
 
 import cudf
 from cudf.testing import assert_eq
+
+
+def test_series_setitem_singleton_range():
+    sr = cudf.Series([1, 2, 3], dtype=np.int64)
+    psr = sr.to_pandas()
+    value = np.asarray([7], dtype=np.int64)
+    sr.iloc[:1] = value
+    psr.iloc[:1] = value
+    assert_eq(sr, cudf.Series([7, 2, 3], dtype=np.int64))
+    assert_eq(sr, psr, check_dtype=True)
 
 
 @pytest.mark.parametrize(
